+++ VersionHistory.md
@@ -1,5 +1,9 @@
 # Version History
 
+### Pending
+
+* Add single parameter `DbParameters.Create` like `DbParameters.Add`.
+
 ### 1.1.0
 
 * Create/add parameters from sequences of parameter tuples with non-`object` value types.